ENH/BF: adjust tutorial for 0.8 change and use of absolute_paths
Main goal is to make tutorial outputs "reproducible".  Consistent output
would allow for automated testing later on, and also would lessen the burden
for a user in attempts to visually match to the paths he/she would be
getting in his/her system.

The only outstanding issue is the shortened path:

    BIDS Layout: .../pybids/bids/tests/data/7t_trt | Subjects: 10 | Sessions: 20 | Runs: 20

if pybids module was "installed" for "development" (e.g. via pip -e .), vs

    BIDS Layout: ...ackages/bids/tests/data/7t_trt | Subjects: 10 | Sessions: 20 | Runs: 20

if installed fully.