ENH: Improve when indexes are forced to be updated
If a requesting a new top-level database file, then trigger
a re-indexing of all the derivatives files as well.

This solution is not a fine-grained as we may want, and it
will re-index in cases when it is perhaps not necessary, but
the overall end-user expectations will be improved (i.e. asking
for an indexing of bids files should result in all files being indexed)