ENH: Remove testing for python 2.7
Resolves: #370.  Starting in January of 2019 discussions for removing
python 2.7 support was discussed.  Agrement was reached that starting
with version 9.0 only python version 3.5 or
greater would be supported.

The py.test invocation is the old version. pytest is the new
version (since 3.0). py.test and pytest invocations will coexist for a
long time, but at some point py.test might be deprecated.