FIX: BIDSLayout -- TypeError: unhashable type: 'dict'
When `BIDSLayout.__repr__` is called on datasets that have `sub-N_scans.tsv` files,
the query to obtain runs also returns these files, and their value is a dictionary
which is not hashable by the `set` object wrapping up the query.

Instead of refining the query (which indeed would fix the issue), this workaround
uses the naive approach of ensuring that each result of the query has `int` value.

Resolves: #681.