Implement automatic detection of derivative data
BIDSLayout now checks if the root dataset is a derivative by looking for
the `DatasetType` entity in `dataset_description.json` to be set to
`derivative`. If it is a derivative, the appropriate config (`==
['bids', 'derivatives']`) will by used for parsing. No check is made if
no valid dataset_description is returned

If validate=True, the derivative dataset is validated using
validate_derivative_paths. Bug in validate_derivative_paths preventing
the validation of dataset_description.json files without either
GeneratedBy or PipelineDescription fields is fixed.

Two new tests are provided:

1) A class of tests checking that:
  a) A dataset with desc entities and PipelineDescription but no
  DatasetType is fully parsed when validate=False, but does not accept
  desc as a filter. It fails to find any files when validate=True
  b) A dataset with DatasetType=derivative is parsed as a derivative.
  desc entities are recognized.
  c) A dataset with DatasetType=derivative but missing both
  PipelineDescription and GeneratedBy raises an error when
  validate=True. If validate=False, it parses as a derivative as in b)

2) Test that db-saved configs not overwritten. This could be a concern
if DatasetType is changed between the times the db is saved and loaded
Add DatasetType=derivative to ds005_derivs test dataset

In test 2, we put the db in a fresh tmpdir rather than the already
existing db-path construct as it was bugging out when multiple tests in
a row used it

Resolves #804