Update indexer to treat .zarr as files not dirs
Microscopy file formats were recently updated in the standard to include
ome-zarr, which are stored in a directory suffixed with '.ome.zarr'.

This adds a simple hardcoded fix to ensure .zarr directories get indexed as files
instead of directories. This enables us to get .ome.zarr microscopy
files, and also prevents needless indexing of .zarr directories that
can contain a huge number of files.