Fix functionality when is_derivative=True and validate=False
Description of the "is_derivative" parameter of BIDSLayout states that this parameter "can be enabled along with validate=False to index derivatives without dataset_description.json." However, the initialization method of BIDSLayout calls "validate_derivative_path(root)" on line 141, which eventually leads to a "FileNotFoundError" as the this function then tries to read a dataset_description.json file that does not exist. This commit fixes this issue by catching this FileNotFoundError and restores intended functionality.