fix for cloud paths
The _check_path_matches_patterns function was failing for cloud (e.g.
s3, gc3) URIs.

 1) the path.relative_to() function doesn't work for cloud URIs
(it just returns the original).
 2) combining Path('/') with the result of relative_to now also throws
an error when using cloud (or non-posix) paths.

This fix drops the uri prefix by using .path to get the posix-like
part of the path, so that the regex check works as expected.