fix: Allow numpy arrays to treat PaddedInt like int
If an exception is raised by `val == self.sval or ...` (that is,
`val.__eq__(self.sval).__bool__()`), we defer to the `val` type to
determine the correct behavior. This is currently only expected to be
triggered by array-like types, but any type that goes this far out of
its way to change its boolean behavior is one we don't want to get too
fancy with.

In the spirit of not getting fancy, we drop any consideration of the
padded string in this context.